+++ src/main/java/walkingkooka/spreadsheet/dominokit/fetcher/FakeConverterFetcherWatcher.java
@@ -20,7 +20,7 @@ package walkingkooka.spreadsheet.dominokit.fetcher;
 import walkingkooka.convert.provider.ConverterInfoSet;
 import walkingkooka.convert.provider.ConverterSelector;
 import walkingkooka.spreadsheet.SpreadsheetId;
-import walkingkooka.spreadsheet.convert.MissingConverter;
+import walkingkooka.spreadsheet.convert.provider.MissingConverter;
 import walkingkooka.spreadsheet.dominokit.AppContext;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
