+++ src/main/java/walkingkooka/spreadsheet/dominokit/pluginaliassetlike/AppContextPluginAliasSetLikeDialogComponentContextConverterAliasSet.java
@@ -29,7 +29,7 @@ import walkingkooka.net.Url;
 import walkingkooka.net.http.HttpMethod;
 import walkingkooka.net.http.HttpStatus;
 import walkingkooka.spreadsheet.SpreadsheetId;
-import walkingkooka.spreadsheet.convert.MissingConverter;
+import walkingkooka.spreadsheet.convert.provider.MissingConverter;
 import walkingkooka.spreadsheet.dominokit.AppContext;
 import walkingkooka.spreadsheet.dominokit.convert.ConverterAliasSetComponent;
 import walkingkooka.spreadsheet.dominokit.fetcher.ConverterFetcherWatcher;
